Refactor command structure by creating separate files for chat, provider, RAG, and system commands, so as to improve organization and maintainability going forward
